Cherry-pick commits from master to sync 0.90-1 branch with production image (#179)
* Bump Python to 3.7.10

* Merge commits from 0.90-1 back to reverted master

* Fix CSV Pipe parsing argument to use weight instead of weights. Fix requirements for tox. (#81)

* Fix script mode training hang with logging enabled. (#77)

* Fix training unit test to match PR #77. (#84)

* Fix label concatenation for RecordIO-protobuf dmatrix (#85)

Closes #83

* Add verbosity to hyperparameter validation. (#87)

* Add verbosity to hyperparameter validation.

* Set scipy requirement to 1.2.2 for sagemaker-containers.

* Add missing eval_metrics to hp validation. (#82)

* Added aucpr and cox-nloglik to eval_metric hp validation.
* Add two separate list for MAXIMIZE and MINIMIZE metrics.

+++ src/sagemaker_xgboost_container/algorithm_mode/serve.py
@@ -176,7 +176,7 @@ def execution_parameters():
         parameters = {
             "MaxConcurrentTransforms": number_of_workers(),
             "BatchStrategy": "MULTI_RECORD",
-            "MaxPayloadInMB": PARSED_MAX_CONTENT_LENGTH / (1024 ** 2)
+            "MaxPayloadInMB": int(PARSED_MAX_CONTENT_LENGTH / (1024 ** 2))
         }
     except Exception as e:
         return flask.Response(response="Unable to determine execution parameters: %s" % e,
